BO-723 Move AvailabilitySummary into a DateTime folder
